# [mutation] Cache tester in AssignMetadata like we do in Assign mutators