Note: depends on #1441 

Per findings in #1441, we can greatly improve the speed of AssignMetadata.Mutate by caching the tester instead of recreating it every time we call Mutate.

This reduces the runtime when no mutation is made by 97%, and when mutations are made by 30%!
```
$ for i in {1..20}; do echo $i; go test github.com/open-policy-agent/gatekeeper/pkg/mutation/mutators --bench=BenchmarkAssignMetadata >> before.out; done
$ git checkout benchmark-metadata-2
$ for i in {1..20}; do echo $i; go test github.com/open-policy-agent/gatekeeper/pkg/mutation/mutators --bench=BenchmarkAssignMetadata >> after.out; done
$ benchstat before.out after.out 

name                        old time/op  new time/op  delta
AssignMetadataMutator_Always-12    6.64µs ± 2%  4.70µs ± 1%  -29.22%  (p=0.000 n=19+19)
AssignMetadataMutator_Never-12     1.84µs ± 6%  0.05µs ± 0%  -97.04%  (p=0.000 n=19+18)
```

I'll send a separate PR which removes the unnecessary json unmarshalling in the case where we perform the mutation, once/if #1439 is merged.